wfm-develop mobile 07.05.2026/wfm/backend/svc_users/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyJWT==2.10.1
cryptography==44.0.0
httpx==0.28.1
semetrics-sdk>=0.1.0


# Dev dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.4
